realtime_sentiment/requirement.txt
opencv-python
opencv-python-headless
pydub
minio
kafka-python
transformers
torch
whisper
fer
deepface
TikTokApi
numpy
imageio
ffmpeg-python
python-dotenv
requests
pytube
moviepy
fastapi
torchaudio
pyspark
mtcnn
uvicorn
yt-dlp
ffmpeg